Introduce separate makefile targets run_client and run_server
The previous "run" target corresponds to the new "run_server".  It
listens for connections on port 1830, so it cannot be run at the same
time as the LINC switch in its default configuration.

The wiki page currently suggests "erl -pa ebin" to start an Erlang shell
to be used as a client.  This is inconvenient, since the ssh
application (and its dependencies) need to be started in order to
connect to a server.