Fix infinite loop in enetconf_parser:other_attributes
It used to go into an infinite loop when getting an <rpc/> element with
an attribute whose name neither is 'message-id' nor begins with 'xmlns'.

I'm not entirely sure what the purpose of this function is, but the
changed version passes the existing unit tests, and also the case I
found the problem with - using a prefix for the netconf base namespace.

(Also allow attributes that start with 'xmlns' but not with 'xmlns:'.)